+++ optimus/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.12.13"
+__version__ = "0.12.14"
 
 
 def safe_commit() -> None: